+++ pages/api/edit-playlist.js
@@ -1,10 +1,17 @@
 import { ApolloServer } from 'apollo-server-micro';
 import { typeDefs } from '@/graphQl/type-definitions';
 import { query } from '@/lib/db';
+import { getSession } from 'next-auth/react';
+import { getUserByEmail } from '@/lib/utils';
 
 const resolvers = {
     Mutation: {
-        editPlaylist: async (root, { input: { id, name, observation, restrictionId, entries } }) => {
+        editPlaylist: async (root, { input: { id, name, observation, restrictionId, entries } }, context) => {
+
+            const user = await getUserByEmail({ context, query });
+            if (!user) {
+                return null;
+            }
 
             if (!id || !name || name.trim.length) {
                 console.error('`id` and `name` are required.');
@@ -16,7 +23,6 @@ const resolvers = {
                 name: name,
                 observation: observation,
                 restrictionId: restrictionId,
-                ownerId: 1,
             }
 
             let results = await query(`
@@ -27,10 +33,22 @@ const resolvers = {
                     observation = ?,
                     restriction_id = ?
                 WHERE 
-                    id = ?
+                    id = ? AND
+                    owner_id = ?
                 `,
-                [playlist.name, playlist.observation, playlist.restrictionId, playlist.id]
+                [
+                    playlist.name,
+                    playlist.observation,
+                    playlist.restrictionId,
+                    playlist.id,
+                    user.id
+                ]
             );
+
+            if (!results) {
+                console.error('User cannot update this playlist.');
+                return null;
+            }
 
             results = await query(`
                 DELETE FROM 
@@ -64,6 +82,13 @@ export const config = {
     },
 }
 
-const server = new ApolloServer({ typeDefs, resolvers });
+const server = new ApolloServer({
+    typeDefs,
+    resolvers,
+    context: async ({ req }) => {
+        const session = await getSession({ req });
+        return { session };
+    },
+});
 
 export default server.createHandler({ path: "/api/edit-playlist" });
